added authentication and permissions

+++ backend/src/resolvers/mutations/account.js
@@ -1,7 +1,9 @@
 import validate from "validate.js";
+import jwt from "jsonwebtoken";
 
 import Account from "model/account";
 import AccountErrors from "resolvers/errors/AccountErrors";
+import PermissionErrors from "resolvers/errors/PermissionErrors";
 
 const constraints = {
   username: {
@@ -22,7 +24,45 @@ const constraints = {
 };
 
 const accountMutations = {
+  async login(parent, args, ctx, info) {
+    const { email, password } = args;
+    // 1. check if there is a user with that email
+    const [account] = await Account.getAccountByEmail(email);
+    if (!account) {
+      throw new AccountErrors.AccountNotFoundError({ data: { email } });
+    }
+    // 2. Check if their password is correct
+    const valid = Account.comparePasswords(password, account.password);
+    if (!valid) {
+      throw new AccountErrors.AccountInvalidPasswordError({ data: { email } });
+    }
+    // 3. generate the JWT Token
+    const token = jwt.sign(
+      { accountId: account.id },
+      process.env.MAILMAN_SECRET
+    );
+    // 4. Set the cookie with the token
+    ctx.response.cookie("token", token, {
+      httpOnly: true,
+      maxAge: 1000 * 60 * 60 * 24 * 365
+    });
+    // 5. Return the user
+    return account;
+  },
+  logout(parent, args, ctx, info) {
+    ctx.response.clearCookie("token");
+    return "Goodbye!";
+  },
   async createAccount(parent, args, ctx, info) {
+    if (!ctx.request.isAdmin) {
+      throw new PermissionErrors.PermissionInsufficient({
+        internalData: {
+          args,
+          info
+        }
+      });
+    }
+
     const {
       username,
       domain,
@@ -63,6 +103,15 @@ const accountMutations = {
   async updateAccount(parent, args, ctx, info) {
     const { id, data } = args;
     const { username, domain, password } = data;
+
+    if (!ctx.request.isAdmin) {
+      throw new PermissionErrors.PermissionInsufficient({
+        internalData: {
+          args,
+          info
+        }
+      });
+    }
 
     // Validation
     const [oldAccount] = await Account.getAccount({ id });
@@ -122,6 +171,15 @@ const accountMutations = {
     }
   },
   async deleteAccount(parent, args, ctx, info) {
+    if (!ctx.request.isAdmin) {
+      throw new PermissionErrors.PermissionInsufficient({
+        internalData: {
+          args,
+          info
+        }
+      });
+    }
+
     const { id } = args;
     await Account.deleteAccount(id);
