+++ backend/src/resolvers/mutations/tlsPolicy.js
@@ -1,8 +1,18 @@
 import TlsPolicy from "model/tlspolicy";
 import TlsPolicyErrors from "resolvers/errors/TlsPolicyErrors";
+import PermissionErrors from "resolvers/errors/PermissionErrors";
 
 const tlsPoliciesMutations = {
   async createTlsPolicy(parent, args, ctx, info) {
+    if (!ctx.request.isAdmin) {
+      throw new PermissionErrors.PermissionInsufficient({
+        internalData: {
+          args,
+          info
+        }
+      });
+    }
+
     const [existsDomain] = await TlsPolicy.getTlsPolicy({
       domain: args.data.domain
     });
@@ -20,6 +30,15 @@ const tlsPoliciesMutations = {
   },
 
   async updateTlsPolicy(parent, args, ctx, info) {
+    if (!ctx.request.isAdmin) {
+      throw new PermissionErrors.PermissionInsufficient({
+        internalData: {
+          args,
+          info
+        }
+      });
+    }
+
     try {
       await TlsPolicy.updateTlsPolicy(args.id, args.data);
       const [tlsPolicy] = await TlsPolicy.getTlsPolicy({ id: args.id });
@@ -30,6 +49,15 @@ const tlsPoliciesMutations = {
   },
 
   async deleteTlsPolicy(parent, args, ctx, info) {
+    if (!ctx.request.isAdmin) {
+      throw new PermissionErrors.PermissionInsufficient({
+        internalData: {
+          args,
+          info
+        }
+      });
+    }
+
     const { id } = args;
     await TlsPolicy.deleteTlsPolicy(id);
